pyproject.toml
[tool.poetry]
name = "marktplaats-gpt"
version = "0.1.0"
description = "ChatGPT for conversations in Marktplaats"
authors = [
  "Aleksandr Vinokurov <aleksandr.vin@gmail.com>"
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/aleksandr-vin/marktplaats-gpt"


[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0"
openai = "^0.28"
marktplaats-messages = "^0.2"
beautifulsoup4 = "^4"
python-telegram-bot = "^20.6"


[tool.poetry.scripts]
"marktplaats-gpt" = "marktplaats_gpt.main:main"
"marktplaats-gpt-bot" = "marktplaats_gpt.bot:main"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/aleksandr-vin/marktplaats-gpt/issues"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-ra --cov=marktplaats_gpt"
testpaths = [
    "tests",
]